# Optimize BulkString writing by combining writes