Speeds up writes for small bulk strings with minimal slow down for larger strings:

```name                                 old time/op  new time/op  delta
WriterWriteBulkStringHeader/0-8      34.7ns ± 1%  24.6ns ± 3%  -29.30%  (p=0.000 n=18+20)
WriterWriteBulkStringHeader/1-8      35.7ns ± 3%  25.5ns ± 3%  -28.73%  (p=0.000 n=20+20)
WriterWriteBulkStringHeader/10-8     36.3ns ± 3%  26.1ns ± 4%  -28.23%  (p=0.000 n=20+20)
WriterWriteBulkStringHeader/100-8    47.7ns ± 4%  37.1ns ± 3%  -22.16%  (p=0.000 n=20+20)
WriterWriteBulkStringHeader/1000-8   60.1ns ± 3%  50.2ns ± 3%  -16.54%  (p=0.000 n=20+20)
WriterWriteBulkStringHeader/10000-8   148ns ± 4%   158ns ± 3%   +6.57%  (p=0.000 n=20+20)
```